New: shorter, more Pythonic method names.
`hhc` is very similar to Python's built in `hex`: hex takes an integer and makes a hexadecimal string while hhc takes an integer and makes a hexahexacontadecimal string. So it makes sense that the two have similar names.

Also the `encode` and `decode` words are used on binary data in Python's built in base64 library, so it would be better to reserve those words for the same in hhc.

This change also adds an __all__ definition.

Since it's not a backwards compatible change, bump the version to 2.0 semver style.