New: backport `.` and `..` avoidance to HHC 2.
This is backwards compatible because `hhc2` won't do it by default, and when it does do it `hhc2_to_int`, both new and old versions, decode to the same number since the padding is just a 0 on the left.

Also fixed using wrong character to pad when a custom alphabet was specified.